cli: split tasks cli tests out of production module
